+++ src/System.Private.CoreLib/shared/Microsoft/Win32/SafeHandles/SafeRegistryHandle.cs
@@ -2,9 +2,14 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using Microsoft.Win32.SafeHandles;
 using System;
 
+#if REGISTRY_ASSEMBLY
 namespace Microsoft.Win32.SafeHandles
+#else
+namespace Internal.Win32.SafeHandles
+#endif
 {
 #if REGISTRY_ASSEMBLY
     public
