+++ src/System.Private.CoreLib/shared/System/Globalization/HijriCalendar.Win32.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Microsoft.Win32;
+using Internal.Win32;
 
 namespace System.Globalization
 {
@@ -47,7 +47,7 @@ namespace System.Globalization
                     return 0;
                 }
 
-                object value = key.GetValue(HijriAdvanceRegKeyEntry, null, RegistryValueOptions.DoNotExpandEnvironmentNames);
+                object value = key.GetValue(HijriAdvanceRegKeyEntry);
                 if (value == null)
                 {
                     return 0;
